backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "fastapi>=0.118.0",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "mysql-connector-python>=9.4.0",
    "pydantic-settings>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.12.0",
    "requests>=2.31.0",
    "uvicorn>=0.37.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
